@@ -3,7 +3,7 @@ package net.n2oapp.platform.jaxrs.api;
 import net.n2oapp.platform.jaxrs.RestCriteria;
 import org.springframework.data.domain.Sort;
 
-import javax.ws.rs.QueryParam;
+import jakarta.ws.rs.QueryParam;
 import java.time.LocalDateTime;
 import java.util.Date;
 import java.util.List;
