@@ -1,7 +1,7 @@
 package net.n2oapp.platform.jaxrs;
 
-import javax.ws.rs.core.Response;
-import javax.ws.rs.ext.ExceptionMapper;
+import jakarta.ws.rs.core.Response;
+import jakarta.ws.rs.ext.ExceptionMapper;
 
 /**
  * Преобразование исключений в ответы REST сервиса
