@@ -2,7 +2,7 @@ package net.n2oapp.platform.selection.integration.model;
 
 import net.n2oapp.platform.selection.api.Selective;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
 
 @Selective
 @MappedSuperclass
