@@ -3,9 +3,9 @@ package net.n2oapp.platform.selection.integration.model;
 import net.n2oapp.platform.selection.api.SelectionIgnore;
 import net.n2oapp.platform.selection.api.Selective;
 
-import javax.persistence.Entity;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
+import jakarta.persistence.Entity;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
 
 @Entity
 @Selective
