@@ -3,7 +3,7 @@ package net.n2oapp.platform.selection.integration.model;
 import net.n2oapp.platform.selection.api.Joined;
 import net.n2oapp.platform.selection.api.Selective;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
 
 @Entity
 @Selective(prefix = "")
