@@ -3,9 +3,9 @@ package net.n2oapp.platform.selection.integration.model;
 import net.n2oapp.platform.selection.api.SelectionIgnore;
 import net.n2oapp.platform.selection.api.Selective;
 
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.ManyToMany;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.ManyToMany;
 import java.util.HashSet;
 import java.util.Set;
 
