@@ -6,7 +6,7 @@ import net.n2oapp.platform.loader.client.auth.ClientContext;
 import net.n2oapp.platform.loader.client.auth.OAuth2ClientContext;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.web.server.LocalServerPort;
+import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.client.RestOperations;
 
